+++ src/components/AddressTab.tsx
@@ -1,17 +1,10 @@
 import { useEffect } from 'react';
-import styled from 'styled-components';
-import Accordion from './Accordion';
-import AttachImage from './AttachImage';
 import useImageUploader from '../hooks/useImageUploader';
+import Accordion from './Accordion';
 import ImageCropper from './ImageCropper';
-import Input from './common/Input';
+import AttachImage from './AttachImage';
 
-const Wrapper = styled.div`
-  display: flex;
-  align-items: center;
-`;
-
-export default function EmailTab() {
+export default function AddressTab() {
   const {
     uploadImage,
     compressedImage,
@@ -24,9 +17,10 @@ export default function EmailTab() {
       handleCompressImage();
     }
   }, [uploadImage]);
+
   return (
-    <Accordion title="이메일">
-      <Wrapper>
+    <Accordion title="주소">
+      <div>
         <ImageCropper aspectRatio={1 / 1} onCrop={handleUploadImage}>
           {compressedImage ? (
             <AttachImage
@@ -39,13 +33,10 @@ export default function EmailTab() {
             <AttachImage width={40} height={40} />
           )}
         </ImageCropper>
-        <Input
-          placeholder="example@email.com"
-          width={257}
-          backgroundColor="#F3F3F3"
-          border="none"
-        />
-      </Wrapper>
+        <input />
+      </div>
+      <input />
+      <input />
     </Accordion>
   );
 }
